Move things out of top-level of lesswrong package
