+++ packages/lesswrong/components/posts/PostsPage/PostsPage.jsx
@@ -5,7 +5,7 @@ import { Posts } from '../../../lib/collections/posts';
 import { Comments } from '../../../lib/collections/comments'
 import { withStyles } from '@material-ui/core/styles';
 import Tooltip from '@material-ui/core/Tooltip';
-import { postBodyStyles } from '../../../themes/stylePiping'
+import { postBodyStyles } from '../../../lib/themes/stylePiping'
 import withUser from '../../common/withUser';
 import withErrorBoundary from '../../common/withErrorBoundary'
 import classNames from 'classnames';
