+++ packages/lesswrong/components/tagging/TagRelCard.jsx
@@ -5,7 +5,7 @@ import { useVote } from '../votes/withVote';
 import { useCurrentUser } from '../common/withUser';
 import { TagRels } from '../../lib/collections/tagRels/collection.js';
 import { Link } from '../../lib/reactRouterWrapper.js';
-import { commentBodyStyles } from '../../themes/stylePiping'
+import { commentBodyStyles } from '../../lib/themes/stylePiping'
 
 const styles = theme => ({
   root: {
