+++ packages/lesswrong/server/material-ui/themeProvider.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { addCallback } from 'meteor/vulcan:core';
 import JssProvider from 'react-jss/lib/JssProvider';
 import { MuiThemeProvider, createGenerateClassName } from '@material-ui/core/styles';
-import forumTheme from '../../themes/forumTheme'
+import forumTheme from '../../lib/themes/forumTheme'
 import { SheetsRegistry } from 'react-jss/lib/jss';
 import JssCleanup from '../../components/themes/JssCleanup';
 import PropTypes from 'prop-types'
